Work around compiler bugs in YukonAtom compiler (GCC 5.2)
> /// .../googlemock/include/gmock/gmock-actions.h:571:15:
> error: 'T testing::DefaultValue<T>::ValueProducer::Produce()
> [with T = testing::{anonymous}::ReturnTest_PrefersConversionOperator_Test::TestBody()::Out]',
> declared using local type 'testing::{anonymous}::ReturnTest_PrefersConversionOperator_Test::TestBody()::Out',
> is used but never defined [-fpermissive]Work around compiler bugs in YukonAtom compiler (GCC 5.2)

> /// .../googlemock/include/gmock/gmock-actions.h:571:15:
> error: 'T testing::DefaultValue<T>::ValueProducer::Produce()
> [with T = testing::{anonymous}::ReturnTest_PrefersConversionOperator_Test::TestBody()::Out]',
> declared using local type 'testing::{anonymous}::ReturnTest_PrefersConversionOperator_Test::TestBody()::Out',
> is used but never defined [-fpermissive]

These appear to be manifestations of a bug fixeed in g++ 5.3, specifically
https://gcc.gnu.org/bugzilla/show_bug.cgi?id=51048

A somewhat minimized reproduction is at https://godbolt.org/z/dxMq9M8o7

The googletest code is valid - the function it's complaining about
is pure virtual and you can't define those (it's always  overridden
in a subclass, or else not called at all). But GCC 5.2 and earlier
error out if the template involved any function-local types.

There's also still a bunch of tests which can't be run because YukonAtom
is stuck on python 3.4, and a very incomplete build of even that
(e.g. it's missing lots of libs that can be copied in from the cpython
v3.4.4 tag, but also built-in things like ./Modules/_posixsubprocess.c

These appear to be manifestations of a bug fixed in g++ 5.3, specifically
https://gcc.gnu.org/bugzilla/show_bug.cgi?id=51048
A somewhat minimized reproduction is at https://godbolt.org/z/dxMq9M8o7

The googletest code is valid - the function it's complaining about
is pure virtual and you can't define those (it's always  overridden
in a subclass, or else not called at all). But GCC 5.2 and earlier
error out if the template involved any function-local types.